+++ tests/Feature/EleTest.php
@@ -5,9 +5,11 @@ namespace Tests\Feature;
 use Tests\TestCase;
 use Illuminate\Foundation\Testing\WithFaker;
 use Illuminate\Foundation\Testing\RefreshDatabase;
+use Illuminate\Foundation\Testing\WithoutMiddleware;
 
 class EleTest extends TestCase
 {
+    // use RefreshDatabase;
     /**
      * A basic feature test example.
      *
@@ -22,12 +24,6 @@ class EleTest extends TestCase
         $response->assertStatus(200);
     }
 
-    // public function Election_register()
-    // {
-    //     // Auth::loginUsingId(2);
-    //     // this->store();
-    // }
-
      /**
      * ログインした状態でリクエストが正しく処理されるか
      * ユーザーを自動で作る
@@ -35,74 +31,106 @@ class EleTest extends TestCase
 
     public function testLoggedIn()
     {
-        $user = factory(\App\User::class)->create();
+        $user = factory(\App\User::class)->make();
  
         $response = $this->actingAs($user)
                          ->get('/');
  
         $response->assertStatus(200);
     }
+    /**
+     * Undocumented function
+     *
+     * @return void
+     * データベースが指定の値を持っているか
+     */
+    public function testDatabase()
+    {
+    // アプリケーションを呼び出す…
 
+    $this->assertDatabaseHas('users', [
+        'email' => 'aa@email.com'
+    ]);
+    }
     /**
      * Undocumented function
      *
      * @return void
      * 選挙を作成して、マイページに飛んだ時、タイトルが表示されるか
      */
-    public function testCreate()
-    {
-        //Ele作る
-        $election = factory('App\Election')->create();
-        // $candidate = factory('App\Candidate')->create();
-        //electionの一覧を表示するURLにアクセスすると
-        $response = $this->get('/elections');
-        
-        //electionが見える
-        $response->assertSee($election->title)
-                ->assertSee($election->subtitle);
 
+    public function testDatabase2()
+    {
+        // 一つのApp\Userインスタンスを作成
+        // $users = factory(\App\User::class)->create()
+        //    ->each(function ($user) {
+        //         $user->elections()->save(factory(\App\Election::class)->make());
+        //     });
+
+              // electionsからリレーションでcandidateを作成
+         $elections = factory(\App\Election::class)->make()
+                ->each(function ($election)
+                { 
+                    $election->candidates()->save(factory(\App\Candidate::class)->make());
+                    });
+         
+                    $this->assertTrue($elections);
+
+                    // candidateからリレーションでelectionはできない？
+        //  $candidate = factory(\App\Candidate::class)->create()
+        //             ->each(function ($candidate)
+        //             { 
+        //                 $candidate->election()->save(factory(\App\Election::class)->make());
+        //                 });
+        // $readElection = $elections->first(); // 取れない
+        // foreach($election as $election)
+        // {
+            //     $this->assertNotNull($election->title); // データが取得できたかテスト
+            // }
+            // $table='elections';
+        // $this->assertDatabaseHas($table, ['title',$readElection->title]);
+        // $this->assertTrue('title', $readElection->title);
+        
+        //electionの一覧を表示するURLにアクセスすると
+        //electionが見える
+        // $response->assertSee($election->title);
     }
+    
+    // }
     /**
      * Undocumented function
      *
      * @return void
      * 
+     * _登録処理が成功する事を検証
      */
-    // public function testCreateElection()
-    // {
-    //     //  electionsからリレーションでcandidateを作成
-    // $elections = factory(\App\Election::class)->create()
-    //      ->each(function ($election)
-    //      { 
-    //          $election->candidates()->save(factory(\App\Candidate::class)->make());
-    //          });
-    //          $elections->assertTrue($elections);
-    // }
-    // public function testBoard()
-    // {
-    // $this->visit('/elections')//  トップページにアクセス
-    // $this->assertSee('みんなの選挙')//      文字列が見える
-    // ->see('選挙作成')//        
-    // ->click('選挙作成')//       リンクをクリックしてみる
-    // ->seePageIs('/elections/new')// 新規投稿ページに遷移する
-    // ->see('選挙登録');//     新規投稿ページには「新規記事投稿」という文字列がある
-    // }
+    
+    public function insert()
+    {
+        $response = $this->get(route('elections.show', ['id' => 19]));
+        $response->assertSee('候補');
+        $response->assertStatus(200);
+    }
 
-    // _登録処理が成功する事を検証
-    // public function insert()
-    // {
-    //     $name = 'test';
-    //     $email = 'test@email.com';
-    //     $pass = 'testtest';
+    public function postVotes()
+    {
+        $response = $this->get(route('elections.show', ['id' => 19]));
 
-    //     $this->users->insert($name, $email, $pass);
-
-    //     $this->assertDatabaseHas('users', [
-    //         'name' => $name,
-    //         'email' => $email,
-    //         'password' => $pass
-    //     ]);
-    // }
+        $this->assertJsonStringEqualsJsonString(
+            json_encode([
+                'id'   => 19,
+                'title' => 'aa',
+            ]),
+            $response->content()
+            );
+            
+            $response = $this->post('/elections/show', [
+                'voted' => 'name1'
+                ]);
+                $this->assertDatabaseHas('votes', [
+                    'voted' => 'name1',
+                    ]);
+   }
 //     public function update_更新処理が成功する事を検証()
 //     {
 //         $id = 1;
